+++ docs/create_new_resource.md
@@ -136,11 +136,7 @@ Your repository should look like this:
 
 namespace App\Repository;
 
-use App\Entity\Book;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-use Sylius\Bundle\ResourceBundle\Doctrine\ORM\ResourceRepositoryTrait;
-use Sylius\Component\Resource\Repository\RepositoryInterface;
+use App\Entity\Book;use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;use Doctrine\Persistence\ManagerRegistry;use Sylius\Bundle\ResourceBundle\Doctrine\ORM\ResourceRepositoryTrait;use Sylius\Resource\Doctrine\Peristence\RepositoryInterface;
 
 /**
  * @extends ServiceEntityRepository<Book>
