+++ src/Bundle/Doctrine/ORM/EntityRepository.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Sylius\Bundle\ResourceBundle\Doctrine\ORM;
 
 use Doctrine\ORM\EntityRepository as BaseEntityRepository;
-use Sylius\Component\Resource\Repository\RepositoryInterface;
+use Sylius\Resource\Doctrine\Peristence\RepositoryInterface;
 
 /** @psalm-suppress DeprecatedInterface */
 class EntityRepository extends BaseEntityRepository implements RepositoryInterface
