+++ src/Bundle/spec/Form/DataTransformer/ResourceToIdentifierTransformerSpec.php
@@ -15,7 +15,7 @@ namespace spec\Sylius\Bundle\ResourceBundle\Form\DataTransformer;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Sylius\Component\Resource\Repository\RepositoryInterface;
+use Sylius\Resource\Doctrine\Peristence\RepositoryInterface;
 use Sylius\Resource\Model\ResourceInterface;
 use Symfony\Component\Form\Exception\TransformationFailedException;
