+++ src/Component/src/Doctrine/Peristence/InMemoryRepository.php
@@ -11,14 +11,14 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Repository;
+namespace Sylius\Resource\Doctrine\Peristence;
 
 use ArrayObject;
 use Pagerfanta\Adapter\ArrayAdapter;
 use Pagerfanta\Pagerfanta;
 use Pagerfanta\PagerfantaInterface;
 use Sylius\Component\Resource\Exception\UnexpectedTypeException;
-use Sylius\Component\Resource\Repository\Exception\ExistingResourceException;
+use Sylius\Component\Resource\src\Doctrine\Peristence\Exception\ExistingResourceException;
 use Sylius\Resource\Model\ResourceInterface;
 use Symfony\Component\PropertyAccess\PropertyAccess;
 use Symfony\Component\PropertyAccess\PropertyAccessor;
