+++ tests/Application/src/Repository/CustomBookRepository.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace App\Repository;
 
 use App\Entity\Book;
-use Sylius\Component\Resource\Repository\RepositoryInterface;
+use Sylius\Resource\Doctrine\Peristence\RepositoryInterface;
 
 final class CustomBookRepository
 {
